@@ -1,14 +1,16 @@
 #ifndef ELLIPSOID_GEOM_GEOLINEARRING_H_
 #define ELLIPSOID_GEOM_GEOLINEARRING_H_
 
-#include <geom/GeoLineString.h>
+#include "geom/GeoLineString.h"
+#include "geom/GeoReferenceSystem.h"
 
 #include <memory>
 
 namespace ep {
 namespace geom {
 
-class GeoLinearRing : public GeoLineString {
+class GeoLinearRing : public GeoLineString
+{
   friend class GeoFactory;
 
 public:
@@ -21,7 +23,8 @@ public:
   bool isClosed() const override;
 
 protected:
-  GeoLinearRing(GeoCoordinateSequence::Ptr &&pts);
+  GeoLinearRing(
+      std::unique_ptr<GeoCoordinateSequence> &&pts, GeoReferenceSystem ref);
 
 private:
   bool validate() const;
