@@ -1,8 +1,9 @@
 #ifndef ELLIPSOID_GEOM_MULTILINESTRING_H_
 #define ELLIPSOID_GEOM_MULTILINESTRING_H_
 
-#include <geom/GeoGeometry.h>
-#include <geom/GeoLineString.h>
+#include "geom/GeoGeometry.h"
+#include "geom/GeoLineString.h"
+#include "geom/GeoReferenceSystem.h"
 
 #include <memory>
 #include <vector>
@@ -10,7 +11,8 @@
 namespace ep {
 namespace geom {
 
-class GeoMultiLineString : public GeoGeometry {
+class GeoMultiLineString : public GeoGeometry
+{
   friend class GeoFactory;
 
 public:
@@ -22,19 +24,21 @@ public:
 
   std::size_t getNumLineString() const;
 
-  GeoLineString *getLineStringN(std::size_t index);
+  GeoLineString       *getLineString(std::size_t index);
+  const GeoLineString *getLineString(std::size_t index) const;
 
-  const GeoLineString *getLineStringN(std::size_t index) const;
+  std::unique_ptr<GeoLineString> getLineStringNew(std::size_t index) const;
 
 #ifdef EP_OPENGL
   void draw() const override;
 #endif
 
 protected:
-  GeoMultiLineString(std::vector<GeoLineString::Ptr> &&lines);
+  GeoMultiLineString(std::vector<std::unique_ptr<GeoLineString>> &&lines,
+      GeoReferenceSystem                                           ref);
 
 private:
-  std::vector<GeoLineString::Ptr> lines_;
+  std::vector<std::unique_ptr<GeoLineString>> lines_;
 };
 
 } // namespace geom
