@@ -1,16 +1,19 @@
 #ifndef ELLIPSOID_GEOM_MULTIPOLYGON_H_
 #define ELLIPSOID_GEOM_MULTIPOLYGON_H_
 
-#include <geom/GeoGeometry.h>
-#include <geom/GeoPolygon.h>
+#include "geom/GeoGeometry.h"
+#include "geom/GeoPolygon.h"
+#include "geom/GeoReferenceSystem.h"
 
+#include <cstddef>
 #include <memory>
 #include <vector>
 
 namespace ep {
 namespace geom {
 
-class GeoMultiPolygon : public GeoGeometry {
+class GeoMultiPolygon : public GeoGeometry
+{
   friend class GeoFactory;
 
 public:
@@ -22,16 +25,18 @@ public:
 
   std::size_t getNumPolygon() const;
 
-  GeoPolygon *getPolygonN(std::size_t index);
+  GeoPolygon       *getPolygon(std::size_t index);
+  const GeoPolygon *getPolygon(std::size_t index) const;
 
-  const GeoPolygon *getPolygonN(std::size_t index) const;
+  std::unique_ptr<GeoPolygon> getPolygonNew(std::size_t index) const;
 
 #ifdef EP_OPENGL
   void draw() const override;
 #endif
 
 protected:
-  GeoMultiPolygon(std::vector<std::unique_ptr<GeoPolygon>> &&ps);
+  GeoMultiPolygon(
+      std::vector<std::unique_ptr<GeoPolygon>> &&ps, GeoReferenceSystem ref);
 
 private:
   std::vector<std::unique_ptr<GeoPolygon>> ps_;
